pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "botfarm-service"
version = "0.1.0"
description = "Async FastAPI botfarm microservice for E2E credentials issuing."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "cryptography>=44.0.0",
  "python-multipart>=0.0.20",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.2",
  "pytest-cov>=6.0.0",
  "httpx>=0.28.1",
  "asgi-lifespan>=2.1.0",
  "aiosqlite>=0.20.0",
  "ruff>=0.9.6",
  "mypy>=1.14.1",
  "pre-commit>=4.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["app", "tests", "alembic"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S", "ASYNC"]
ignore = ["S104", "S105", "S106", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config --cov=app --cov-report=term-missing --cov-fail-under=75"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true
plugins = []

[tool.hatch.build.targets.wheel]
packages = ["app"]
